+++ mimic-bukkit/src/test/kotlin/impl/battlelevels/BattleLevelsLevelSystemTest.kt
@@ -19,184 +19,130 @@
 
 package ru.endlesscode.mimic.impl.battlelevels
 
-import com.nhaarman.mockitokotlin2.*
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.verify
 import ru.endlesscode.mimic.BukkitTestBase
-import kotlin.test.BeforeTest
 import kotlin.test.Test
 
 class BattleLevelsLevelSystemTest : BukkitTestBase() {
 
-    private lateinit var battleLevelsApi: BattleLevelsApiWrapper
+    private val battleLevelsApi: BattleLevelsApiWrapper = mockk(relaxUnitFun = true) {
+        // Our formula exp to next level is static = 10
+        every { getNeededFor(any()) } answers { arg<Int>(0) * 10.0 }
+    }
 
     // SUT
-    private lateinit var levelSystem: BattleLevelsLevelSystem
-
-    @BeforeTest
-    override fun setUp() {
-        super.setUp()
-        battleLevelsApi = mock {
-            // Our formula exp to next level is static = 10
-            on { getNeededFor(any()) } doAnswer { it.getArgument<Int>(0) * 10.0 }
-        }
-        levelSystem = BattleLevelsLevelSystem(player, battleLevelsApi)
-    }
+    private val levelSystem: BattleLevelsLevelSystem = BattleLevelsLevelSystem(player, battleLevelsApi)
 
     @Test
     fun `when set level lower than current - should remove level`() {
-        // Given
         set(level = 10)
-
-        // When
         levelSystem.level = 6
 
-        // Then
-        verify(battleLevelsApi).removeLevel(any(), eq(4))
+        verify { battleLevelsApi.removeLevel(any(), eq(4)) }
     }
 
     @Test
     fun `when set level higher than current - should add level`() {
-        // Given
         set(level = 10)
-
-        // When
         levelSystem.level = 16
 
-        // Then
-        verify(battleLevelsApi).addLevel(any(), eq(6))
+        verify { battleLevelsApi.addLevel(any(), eq(6)) }
     }
 
     @Test
     fun `when set level equal to current - should not change level`() {
-        // Given
         set(level = 10)
-
-        // When
         levelSystem.level = 10
 
-        // Then
-        verify(battleLevelsApi, never()).removeLevel(any(), any())
-        verify(battleLevelsApi, never()).addLevel(any(), any())
+        verify(exactly = 0) { battleLevelsApi.removeLevel(any(), any()) }
+        verify(exactly = 0) { battleLevelsApi.addLevel(any(), any()) }
     }
 
     @Test
     fun `when set total exp lower than current - should remove score`() {
-        // Given
-        set(totalExp = 100.0)
-
-        // When
+        set(level = 1, totalExp = 100.0)
         levelSystem.totalExp = 60.0
 
-        // Then
-        verify(battleLevelsApi).removeScore(any(), eq(40.0))
+        verify { battleLevelsApi.removeScore(any(), eq(40.0)) }
     }
 
     @Test
     fun `when set total exp higher than current - should add score`() {
-        // Given
         set(level = 10, totalExp = 100.0)
-
-        // When
         levelSystem.totalExp = 160.0
 
-        // Then
-        verify(battleLevelsApi).addLevel(any(), eq(6))
+        verify { battleLevelsApi.addLevel(any(), eq(6)) }
     }
 
     @Test
     fun `when set total exp equal to current - should not change score`() {
-        // Given
         set(totalExp = 100.0)
-
-        // When
         levelSystem.totalExp = 100.0
 
-        // Then
-        verify(battleLevelsApi, never()).removeScore(any(), any())
-        verify(battleLevelsApi, never()).addScore(any(), any())
+        verify(exactly = 0) { battleLevelsApi.removeScore(any(), any()) }
+        verify(exactly = 0) { battleLevelsApi.addScore(any(), any()) }
     }
 
     @Test
     fun `when give exp more than one level - should add level and give extra exp`() {
-        // Given
         set(level = 1, totalExp = 16.0)
-
-        // When
         levelSystem.giveExp(28.0)
 
-        // Then
-        verify(battleLevelsApi).addLevel(any(), eq(3))
-        verify(battleLevelsApi).addScore(any(), eq(4.0))
+        verify { battleLevelsApi.addLevel(any(), eq(3)) }
+        verify { battleLevelsApi.addScore(any(), eq(4.0)) }
     }
 
     @Test
     fun `when give exp equal to one level - should add level`() {
-        // Given
         set(level = 1, totalExp = 16.0)
-
-        // When
         levelSystem.giveExp(4.0)
 
-        // Then
-        verify(battleLevelsApi).addLevel(any(), eq(1))
-        verify(battleLevelsApi, never()).addScore(any(), any())
+        verify { battleLevelsApi.addLevel(any(), eq(1)) }
+        verify(exactly = 0) { battleLevelsApi.addScore(any(), any()) }
     }
 
     @Test
     fun `when give exp not more than one level - should add exp`() {
-        // Given
         set(level = 1, totalExp = 16.0)
-
-        // When
         levelSystem.giveExp(3.0)
 
-        // Then
-        verify(battleLevelsApi).addScore(any(), eq(3.0))
-        verify(battleLevelsApi, never()).addLevel(any(), any())
+        verify { battleLevelsApi.addScore(any(), eq(3.0)) }
+        verify(exactly = 0) { battleLevelsApi.addLevel(any(), any()) }
     }
 
     @Test
     fun `when take exp more than current exp - should take level and give extra exp`() {
-        // Given
         set(level = 10, totalExp = 106.0)
-
-        // When
         levelSystem.takeExp(28.0)
 
-        // Then
-        verify(battleLevelsApi).removeLevel(any(), eq(3))
-        verify(battleLevelsApi).addScore(any(), eq(8.0))
-        verify(battleLevelsApi, never()).removeScore(any(), any())
+        verify { battleLevelsApi.removeLevel(any(), eq(3)) }
+        verify { battleLevelsApi.addScore(any(), eq(8.0)) }
+        verify(exactly = 0) { battleLevelsApi.removeScore(any(), any()) }
     }
 
     @Test
     fun `when take exp equal to two levels - should take level and not give extra exp`() {
-        // Given
         set(level = 10, totalExp = 106.0)
-
-        // When
         levelSystem.takeExp(26.0)
 
-        // Then
-        verify(battleLevelsApi).removeLevel(any(), eq(2))
-        verify(battleLevelsApi, never()).addScore(any(), any())
-        verify(battleLevelsApi, never()).removeScore(any(), any())
+        verify { battleLevelsApi.removeLevel(any(), eq(2)) }
+        verify(exactly = 0) { battleLevelsApi.addScore(any(), any()) }
+        verify(exactly = 0) { battleLevelsApi.removeScore(any(), any()) }
     }
 
     @Test
     fun `when take exp not more than current exp - should take only exp`() {
-        // Given
         set(level = 10, totalExp = 106.0)
-
-        // When
         levelSystem.takeExp(6.0)
 
-        // Then
-        verify(battleLevelsApi).removeScore(any(), eq(6.0))
-        verify(battleLevelsApi, never()).removeLevel(any(), any())
+        verify { battleLevelsApi.removeScore(any(), eq(6.0)) }
+        verify(exactly = 0) { battleLevelsApi.removeLevel(any(), any()) }
     }
 
     private fun set(level: Int? = null, totalExp: Double? = null) {
-        if (level != null) whenever(battleLevelsApi.getLevel(any())) doReturn level
-        if (totalExp != null) whenever(battleLevelsApi.getScore(any())) doReturn totalExp
+        if (level != null) every { battleLevelsApi.getLevel(any()) } returns level
+        if (totalExp != null) every { battleLevelsApi.getScore(any()) } returns totalExp
     }
 }