+++ mimic-bukkit/src/test/kotlin/impl/skillapi/SkillApiLevelSystemTest.kt
@@ -19,149 +19,90 @@
 
 package ru.endlesscode.mimic.impl.skillapi
 
-import com.nhaarman.mockitokotlin2.doReturn
-import com.nhaarman.mockitokotlin2.verify
-import com.nhaarman.mockitokotlin2.whenever
 import com.sucy.skill.api.enums.ExpSource
 import com.sucy.skill.api.player.PlayerClass
+import io.kotest.matchers.shouldBe
+import io.mockk.every
+import io.mockk.verify
 import ru.endlesscode.mimic.level.LevelSystem
-import kotlin.test.BeforeTest
 import kotlin.test.Test
-import kotlin.test.assertEquals
 
 @Suppress("DEPRECATION")
 class SkillApiLevelSystemTest : SkillApiTestBase() {
 
     // SUT
-    private lateinit var levelSystem: LevelSystem
+    private val levelSystem: LevelSystem = SkillApiLevelSystem(player, skillApi)
 
-    private lateinit var playerClass: PlayerClass
+    private val playerClass: PlayerClass
 
-    @BeforeTest
-    override fun setUp() {
-        super.setUp()
-        levelSystem = SkillApiLevelSystem(player, skillApi)
+    init {
+        prepareClasses("Primary")
+        playerClass = data.mainClass
     }
 
     @Test
     fun `when decrease level - should set right level`() {
-        // Given
-        prepareMainClass()
-        whenever(levelSystem.level) doReturn 3
-
-        // When
+        set(level = 3, exp = 0.0)
         levelSystem.takeLevels(1)
 
-        // Then
-        verify(playerClass).level = 2
+        verify { playerClass.level = 2 }
     }
 
     @Test
     fun `when increase level - should call giveLevels`() {
-        // Given
-        prepareMainClass()
-        val lvlAmount = 10
-
-        // When
-        levelSystem.giveLevels(lvlAmount)
-
-        // Then
-        verify(playerClass).giveLevels(lvlAmount)
+        levelSystem.giveLevels(10)
+        verify { playerClass.giveLevels(10) }
     }
 
     @Test
     fun `when get level - should get class level`() {
-        // Given
-        prepareMainClass()
-
-        // When
-        levelSystem.level
-
-        // Then
-        verify(playerClass).level
+        set(level = 10)
+        playerClass.level shouldBe 10
     }
 
     @Test
     fun `when set level - should change level of class`() {
-        // Given
-        prepareMainClass()
-        val newLevel = 10
-
-        // When
-        levelSystem.level = newLevel
-
-        // Then
-        verify(playerClass).level = newLevel
+        levelSystem.level = 10
+        verify { playerClass.level = 10 }
     }
 
     @Test
     fun `when take exp - should call loseExp with percents`() {
-        // Given
-        prepareMainClass()
-        whenever(playerClass.requiredExp) doReturn 10
-
-        // When
+        set(requiredExp = 10)
         levelSystem.takeExp(5.0)
 
-        // Then
-        verify(playerClass).loseExp(0.5)
+        verify { playerClass.loseExp(0.5) }
     }
 
     @Test
     fun `when give exp - should call give exp with right exp and special source`() {
-        // Given
-        prepareMainClass()
-        val expAmount = 50.0
-
-        // When
-        levelSystem.giveExp(expAmount)
-
-        // Then
-        verify(playerClass).giveExp(expAmount, ExpSource.SPECIAL)
+        levelSystem.giveExp(50.0)
+        verify { playerClass.giveExp(50.0, ExpSource.SPECIAL) }
     }
 
     @Test
     fun `when get exp - should return exp of class`() {
-        // Given
-        prepareMainClass()
-
-        // When
-        levelSystem.exp
-
-        // Then
-        verify(playerClass).exp
+        set(exp = 42.0)
+        levelSystem.exp shouldBe 42.0
     }
 
     @Test
     fun `when set exp - should set class exp`() {
-        // Given
-        prepareMainClass()
-        whenever(playerClass.level) doReturn 2
-        val expAmount = 10.0
+        set(level = 2)
 
-        // When
-        levelSystem.exp = expAmount
-
-        // Then
-        verify(playerClass).exp = expAmount
+        levelSystem.exp = 10.0
+        verify { playerClass.exp = 10.0 }
     }
 
     @Test
     fun `when get exp to next level - should return right exp`() {
-        // Given
-        prepareMainClass()
-        whenever(playerClass.exp) doReturn 70.0
-        whenever(playerClass.level) doReturn 10
-
-        // When
-        val actual = levelSystem.expToNextLevel
-
-        // Then
-        assertEquals(30.0, actual)
+        set(level = 10, exp = 70.0)
+        levelSystem.expToNextLevel shouldBe 30.0
     }
 
-    private fun prepareMainClass() {
-        prepareClasses("Primary")
-        playerClass = data.mainClass
+    private fun set(level: Int? = null, exp: Double? = null, requiredExp: Int? = null) {
+        if (level != null) every { playerClass.level } returns level
+        if (exp != null) every { playerClass.exp } returns exp
+        if (requiredExp != null) every { playerClass.requiredExp } returns requiredExp
     }
 }